# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.StoreDisciplineDTO:
    properties:
      assessment_type:
        example: exam
        type: string
      code:
        example: "00342342413"
        type: string
      competencies:
        example: 1,2,3
        type: string
      desc:
        example: John Doe
        type: string
      hours_individual:
        example: 0
        type: integer
      hours_internship:
        example: 0
        type: integer
      hours_practice:
        example: 50
        type: integer
      hours_self_study:
        example: 0
        type: integer
      hours_theory:
        example: 50
        type: integer
      hours_total:
        example: 100
        type: integer
      hours_with_teacher:
        example: 0
        type: integer
      lang:
        example: en
        type: string
      name:
        example: Discipline Name
        type: string
      necessities:
        example: 1,2,3
        type: string
      post_requisites:
        example: 1,2,3
        type: string
      pre_requisites:
        example: 1,2,3
        type: string
      qualification_id:
        example: 1
        type: integer
    required:
    - name
    - qualification_id
    type: object
  dto.StoreDisciplineModuleChapterDTO:
    properties:
      discipline_module_id:
        example: 1
        type: integer
      name:
        example: DisciplineModuleChapter Name
        type: string
    required:
    - discipline_module_id
    - name
    type: object
  dto.StoreDisciplineModuleChapterTopicDTO:
    properties:
      discipline_module_chapter_id:
        example: 1
        type: integer
      hours_individual:
        example: 0
        type: integer
      hours_internship:
        example: 0
        type: integer
      hours_practice:
        example: 50
        type: integer
      hours_self_study:
        example: 0
        type: integer
      hours_theory:
        example: 50
        type: integer
      hours_with_teacher:
        example: 0
        type: integer
      name:
        example: DisciplineModuleChapterTopic Name
        type: string
      type:
        enum:
        - practice
        - theory
        - combined
        example: practice
        type: string
    required:
    - discipline_module_chapter_id
    - name
    - type
    type: object
  dto.StoreDisciplineModuleDTO:
    properties:
      discipline_id:
        example: 1
        type: integer
      name:
        example: DisciplineModule Name
        type: string
    required:
    - discipline_id
    - name
    type: object
...
  entity.Discipline:
    properties:
      assessment_type:
        example: exam
        type: string
      code:
        example: JHN
        type: string
      competencies:
        example: 1,2,3
        type: string
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      desc:
        example: John Doe
        type: string
      discipline_modules:
        items:
          $ref: '#/definitions/entity.DisciplineModule'
        type: array
      hours_individual:
        example: 0
        type: integer
      hours_internship:
        example: 0
        type: integer
      hours_practice:
        example: 50
        type: integer
      hours_self_study:
        example: 0
        type: integer
      hours_theory:
        example: 50
        type: integer
      hours_total:
        example: 100
        type: integer
      hours_with_teacher:
        example: 0
        type: integer
      id:
        example: 1
        type: integer
      lang:
        example: en
        type: string
      name:
        example: John
        type: string
      necessities:
        example: 1,2,3
        type: string
      post_requisites:
        example: 1,2,3
        type: string
      pre_requisites:
        example: 1,2,3
        type: string
      qualification_id:
        example: 1
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  entity.DisciplineModule:
    properties:
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      discipline_id:
        example: 1
        type: integer
      discipline_module_chapters:
        items:
          $ref: '#/definitions/entity.DisciplineModuleChapter'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: John
        type: string
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
  entity.DisciplineModuleChapter:
    properties:
      code:
        example: JHN
        type: string
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      discipline_module_chapter_topics:
        items:
          $ref: '#/definitions/entity.DisciplineModuleChapterTopic'
        type: array
      discipline_module_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: John
        type: string
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
  entity.DisciplineModuleChapterTopic:
    properties:
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      discipline_module_chapter_id:
        example: 1
        type: integer
      hours_individual:
        example: 0
        type: integer
      hours_internship:
        example: 0
        type: integer
      hours_practice:
        example: 50
        type: integer
      hours_self_study:
        example: 0
        type: integer
      hours_theory:
        example: 50
        type: integer
      hours_with_teacher:
        example: 0
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: John
        type: string
      type:
        example: practice
        type: string
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
...
  v1.getDisciplineModuleChapterTopicsResponse:
    properties:
      disciplineModuleChapterTopics:
        items:
          $ref: '#/definitions/entity.DisciplineModuleChapterTopic'
        type: array
      total:
        type: integer
    type: object
  v1.getDisciplineModuleChaptersResponse:
    properties:
      disciplineModuleChapters:
        items:
          $ref: '#/definitions/entity.DisciplineModuleChapter'
        type: array
      total:
        type: integer
    type: object
  v1.getDisciplineModulesResponse:
    properties:
      disciplineModules:
        items:
          $ref: '#/definitions/entity.DisciplineModule'
        type: array
      total:
        type: integer
    type: object
  v1.getDisciplinesResponse:
    properties:
      disciplines:
        items:
          $ref: '#/definitions/entity.Discipline'
        type: array
      total:
        type: integer
    type: object
...
  v1.response:
    properties:
      message:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /v1/disciplineModuleChapterTopics:
    get:
      consumes:
      - application/json
      description: Fetch a list of disciplineModuleChapterTopics.
      operationId: get_disciplineModuleChapterTopics
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with disciplineModuleChapterTopic list
          schema:
            $ref: '#/definitions/v1.getDisciplineModuleChapterTopicsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get DisciplineModuleChapterTopics
      tags:
      - DisciplineModuleChapterTopics
    post:
      consumes:
      - application/json
      description: Store a new disciplineModuleChapterTopic.
      operationId: store_disciplineModuleChapterTopic
      parameters:
      - description: DisciplineModuleChapterTopic request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StoreDisciplineModuleChapterTopicDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully stored
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Store DisciplineModuleChapterTopic
      tags:
      - DisciplineModuleChapterTopics
  /v1/disciplineModuleChapterTopics/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a disciplineModuleChapterTopic.
      operationId: delete_disciplineModuleChapterTopic
      parameters:
      - description: DisciplineModuleChapterTopic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete DisciplineModuleChapterTopic
      tags:
      - DisciplineModuleChapterTopics
    get:
      consumes:
      - application/json
      description: Fetch a disciplineModuleChapterTopic by ID.
      operationId: get_disciplineModuleChapterTopic
      parameters:
      - description: DisciplineModuleChapterTopic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with disciplineModuleChapterTopic
          schema:
            $ref: '#/definitions/entity.DisciplineModuleChapterTopic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get DisciplineModuleChapterTopic
      tags:
      - DisciplineModuleChapterTopics
  /v1/disciplineModuleChapters:
    get:
      consumes:
      - application/json
      description: Fetch a list of disciplineModuleChapters.
      operationId: get_disciplineModuleChapters
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with disciplineModuleChapter list
          schema:
            $ref: '#/definitions/v1.getDisciplineModuleChaptersResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get DisciplineModuleChapters
      tags:
      - DisciplineModuleChapters
    post:
      consumes:
      - application/json
      description: Store a new disciplineModuleChapter.
      operationId: store_disciplineModuleChapter
      parameters:
      - description: DisciplineModuleChapter request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StoreDisciplineModuleChapterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully stored
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Store DisciplineModuleChapter
      tags:
      - DisciplineModuleChapters
  /v1/disciplineModuleChapters/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a disciplineModuleChapter.
      operationId: delete_disciplineModuleChapter
      parameters:
      - description: DisciplineModuleChapter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete DisciplineModuleChapter
      tags:
      - DisciplineModuleChapters
    get:
      consumes:
      - application/json
      description: Fetch a disciplineModuleChapter by ID.
      operationId: get_disciplineModuleChapter
      parameters:
      - description: DisciplineModuleChapter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with disciplineModuleChapter
          schema:
            $ref: '#/definitions/entity.DisciplineModuleChapter'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get DisciplineModuleChapter
      tags:
      - DisciplineModuleChapters
  /v1/disciplineModules:
    get:
      consumes:
      - application/json
      description: Fetch a list of disciplineModules.
      operationId: get_disciplineModules
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with disciplineModule list
          schema:
            $ref: '#/definitions/v1.getDisciplineModulesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get DisciplineModules
      tags:
      - DisciplineModules
    post:
      consumes:
      - application/json
      description: Store a new disciplineModule.
      operationId: store_disciplineModule
      parameters:
      - description: DisciplineModule request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StoreDisciplineModuleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully stored
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Store DisciplineModule
      tags:
      - DisciplineModules
  /v1/disciplineModules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a disciplineModule.
      operationId: delete_disciplineModule
      parameters:
      - description: DisciplineModule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete DisciplineModule
      tags:
      - DisciplineModules
    get:
      consumes:
      - application/json
      description: Fetch a disciplineModule by ID.
      operationId: get_disciplineModule
      parameters:
      - description: DisciplineModule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with disciplineModule
          schema:
            $ref: '#/definitions/entity.DisciplineModule'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get DisciplineModule
      tags:
      - DisciplineModules
  /v1/disciplines:
    get:
      consumes:
      - application/json
      description: Fetch a list of disciplines.
      operationId: get_disciplines
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with discipline list
          schema:
            $ref: '#/definitions/v1.getDisciplinesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get Disciplines
      tags:
      - Disciplines
    post:
      consumes:
      - application/json
      description: Store a new discipline.
      operationId: store_discipline
      parameters:
      - description: Discipline request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StoreDisciplineDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully stored
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Store Discipline
      tags:
      - Disciplines
  /v1/disciplines/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a discipline.
      operationId: delete_discipline
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete Discipline
      tags:
      - Disciplines
    get:
      consumes:
      - application/json
      description: Fetch a discipline by ID.
      operationId: get_discipline
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with discipline
          schema:
            $ref: '#/definitions/entity.Discipline'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get Discipline
      tags:
      - Disciplines
...
swagger: "2.0"